teams/requirements.txt
asgiref==3.4.1
Django==3.2.5
django-admin-rangefilter==0.8.1
djangorestframework==3.12.4
drf-writable-nested==0.6.3
mailer==0.8.1
Pillow==8.3.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
